Player Instantiation Fix
Player now instantiates on the ServerScene, then the player is moved to its actual scene on the server.
The players layer is also set to the same layer as scene name, and all gameObjects in every scene must sit on the same layer as the scene name, so physics ONLY works for gameObjects on those scenes. In project setting, I only enabled physics between the same layers only, so two objects in different layers cannot interact.